[Feature] Bind taks with users

+++ todos/views.py
@@ -7,7 +7,7 @@ from django.contrib.auth.decorators import login_required
 
 @login_required
 def index(request):
-    tasks = Task.objects.all().order_by("created_at")
+    tasks = Task.objects.filter(user=request.user).order_by("created_at")
 
     return render(request, "tasks/index.html", {"tasks": tasks})
 
@@ -25,9 +25,10 @@ def create(request):
 def update(request, pk):
     try:
         task = Task.objects.get(id=pk)
-        form = TaskForm(instance=task)
-        if request.method == "POST":
-            return _handle_form_update(req=request, instance=task)
+        if task.user == request.user:
+            form = TaskForm(instance=task)
+            if request.method == "POST":
+                return _handle_form_update(req=request, instance=task)
 
         return render(request, "tasks/update.html", {"task_edit_form": form})
 
@@ -40,7 +41,7 @@ def update(request, pk):
 def delete(request, pk):
     try:
         task = Task.objects.get(id=pk)
-        task.delete()
+        task.delete() if task.user == request.user else None
         return redirect(TasksUrls.INDEX.value)
 
     except:
@@ -51,5 +52,7 @@ def delete(request, pk):
 def _handle_form_update(req, instance=None):
     form = TaskForm(req.POST) if instance is None else TaskForm(req.POST, instance=instance)
     if form.is_valid():
-        form.save()
+        task = form.save(commit=False)
+        task.user = req.user
+        task.save()
         return redirect(TasksUrls.INDEX.value)
